--- tests/e2e/suites/basic/test_data/hello-world_values.yaml
@@ -1,12 +1,0 @@
-clusterName: {{ .ClusterName }}
-organization: {{ .Organization }}
-ingress:
-  hosts:
-    - host: {{ index .ExtraValues "IngressHost" }}
-      paths:
-        - path: /
-          pathType: Prefix
-  tls:
-    - secretName: hello-world-tls
-      hosts:
-        - {{ index .ExtraValues "IngressHost" }}
