--- tests/e2e/suites/basic/test_data/nginx_values.yaml
@@ -1,5 +1,0 @@
-controller:
-  service:
-    loadBalancerClass: service.k8s.aws/nlb
-    annotations:
-      service.beta.kubernetes.io/aws-load-balancer-scheme: internet-facing
